Unite onesweep and onesweep_by_key headers
Move onesweep kernels into a separate header
